fixes gitlab issues
