+++ deploy/services/cli/dtaas_services/templates/config/services.env.template
@@ -1,12 +1,12 @@
 # Common settings
-HOSTNAME=services.foo.com
+HOSTNAME=services.intocps.org
 USERID=0
 GROUPID=0
 
 # Certs path on Linux / MacOS
-CERTS_SRC='/etc/letsencrypt/archive/services.foo.com'
+CERTS_SRC='/etc/letsencrypt/archive/services.intocps.org'
 # Certs path on Windows
-#CERTS_SRC='C:/Certbot/archive/services.foo.com'
+#CERTS_SRC='C:/Certbot/archive/services.intocps.org'
 
 # RabbitMQ settings
 RABBITMQ_PORT=8083
